+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version {next}
+  - Better tag normalization code organisation
+
 ## Version 0.0.2 (2016-05-25)
   - Fix issue#1: tagged notes page needs the tag from the URL normalizing.